% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental_functions.R
\name{format_p_val}
\alias{format_p_val}
\title{Format p-value with a specified significant digit cutoff}
\usage{
format_p_val(x, digits = 3)
}
\arguments{
\item{x}{A vector of p-values}

\item{digits}{how many significant digits are to be used.}
}
\value{
A vector of formatted p-values.
}
\description{
\code{format_p_value} will take a set of p-values, round to the specified
significant digits, and convert values that are less than the specified
cutoff to "<'cutoff'"
}
