% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental_functions.R
\name{gather_factors}
\alias{gather_factors}
\title{Combine the levels or unique values of multiple columns into a vector}
\usage{
gather_factors(x, cols, rev = FALSE)
}
\arguments{
\item{x}{A dataframe}

\item{cols}{The columns from which to grab unique values and factor levels}

\item{rev}{Specifies which columns will have a reversed order}
}
\value{
Returns a vector of unique values and levels ordered in a manner
consistent with the columns from which they were derived.
}
\description{
\code{gather_factors} is a function designed to collect the factor levels
and unique values within a series of columns and combine them into one
vector. This may be useful when trying to construct a table using
\code{htmlTable_td}.
}
