% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental_functions.R
\name{get_first_levels}
\alias{get_first_levels}
\title{Function to retrieve the first level from one or more columns in a data.frame}
\usage{
get_first_levels(x, cols)
}
\arguments{
\item{x}{A dataframe}

\item{cols}{The columns from which to retrieve first levels}
}
\value{
Returns a vector of first levels
}
\description{
\code{get_first_levels} is a function that may be used to retrieve the first
level from one or more columns in a data.frame
}
