% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_documentations.R
\docType{package}
\name{tidyhTbl}
\alias{tidyhTbl}
\alias{tidyhTbl-package}
\title{tidyhTbl: A tidy data implementation of the htmlTable package}
\description{
This package contains wrapper functions for the htmlTable package and
accessory functions to help format tidy data.
}
\section{htmlTable wrappers}{

htmlTable_td
}

\section{Accessory functions}{

get_factor_levels
get_first_levels
}

