% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_td.R
\name{htmlTable_td}
\alias{htmlTable_td}
\title{Generate an htmlTable from tidy data}
\usage{
htmlTable_td(x, value = "value", header_td = "header",
  rnames_td = "rnames", rgroup_td = NULL, hidden_rgroup = NULL,
  cgroup1_td = NULL, cgroup2_td = NULL, tspanner_td = NULL,
  hidden_tspanner = NULL, ...)
}
\arguments{
\item{x}{The tidy data used to build the table}

\item{value}{The individual values which will fill each cell of the
table}

\item{header_td}{The column in \code{x} specifying column headings}

\item{rnames_td}{The column in \code{x} specifying row names}

\item{rgroup_td}{The column in \code{x} specifying row groups}

\item{cgroup1_td}{The column in \code{x} specifying the inner most column
groups}

\item{cgroup2_td}{The column in \code{x} specifying the outer most column
groups}

\item{tspanner_td}{The column in \code{x} specifying tspanner groups}
}
\value{
Returns html code that will build a pretty table.
}
\description{
A wrapper script for \code{htmlTable} that will build an htmlTable using
tidy data and mapping elements of the table to specific columns.
}
\seealso{
\code{\link[htmlTable]{htmlTable}}
}
