% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_td.R
\name{htmlTable_td}
\alias{htmlTable_td}
\title{Generate an htmlTable from tidy data}
\usage{
htmlTable_td(x, value = "value", header_td = "header",
  rnames_td = "rnames", rgroup_td = NULL, hidden_rgroup = NULL,
  cgroup1_td = NULL, cgroup2_td = NULL, tspanner_td = NULL,
  hidden_tspanner = NULL, debug_l = FALSE, ...)
}
\arguments{
\item{x}{The tidy data used to build the table}

\item{value}{The individual values which will fill each cell of the
table}

\item{header_td}{The column in \code{x} specifying column headings}

\item{rnames_td}{The column in \code{x} specifying row names}

\item{rgroup_td}{The column in \code{x} specifying row groups}

\item{hidden_rgroup}{rgroups that will be hidden.}

\item{cgroup1_td}{The column in \code{x} specifying the inner most column
groups}

\item{cgroup2_td}{The column in \code{x} specifying the outer most column
groups}

\item{tspanner_td}{The column in \code{x} specifying tspanner groups}

\item{hidden_tspanner}{tspanners that will be hidden.}

\item{debug_l}{Setting to true will print a message at certain points in the
execution of the function. This is meant to only be a temporary feature
as I learn to better utilizing existing debugging features in R Studio.}

\item{...}{Additional arguments that will be passed to the inner
\code{htmlTable} function}
}
\value{
Returns html code that will build a pretty table
}
\description{
A wrapper script for \code{htmlTable} that will build an htmlTable using
tidy data and mapping elements of the table to specific columns.
}
\examples{
\dontrun{
library(tidyverse)
mtcars \%>\%
    rownames_to_column \%>\%
    select(rowname, cyl, gear, hp, mpg, qsec) \%>\%
    gather(per_metric, value, hp, mpg, qsec) \%>\%
    group_by(cyl, gear, per_metric) \%>\%
    summarise(Mean = round(mean(value), 1),
              SD = round(sd(value), 1),
              Min = round(min(value), 1),
              Max = round(max(value), 1)) \%>\%
     gather(summary_stat, value, Mean, SD, Min, Max) \%>\%
     ungroup \%>\%
     mutate(gear = paste(gear, "Gears"),
            cyl = paste(cyl, "Cylinders")) \%>\%
     htmlTable_td(header_td = "gear",
                  cgroup1_td = "cyl",
                  cell_value = "value",
                  rnames_td = "summary_stat",
                  rgroup_td = "per_metric")
}
}
\seealso{
\code{\link[htmlTable]{htmlTable}}
}
